Move PageAllocatedVector into the clp namespace.
