+++ components/core/src/clp/clg/CMakeLists.txt
@@ -28,6 +28,7 @@ set(
         ../MySQLParamBindings.hpp
         ../MySQLPreparedStatement.cpp
         ../MySQLPreparedStatement.hpp
+        ../PageAllocatedVector.hpp
         ../ParsedMessage.cpp
         ../ParsedMessage.hpp
         ../Query.cpp
@@ -92,7 +93,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.hpp"
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.inc"
         "${PROJECT_SOURCE_DIR}/src/ir/types.hpp"
-        "${PROJECT_SOURCE_DIR}/src/PageAllocatedVector.hpp"
         "${PROJECT_SOURCE_DIR}/src/Platform.hpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.cpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.hpp"
